moves choose and crop to templates module
