+++ src/yass/templates/crop.py
